add-most-popular

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -20,7 +20,7 @@ public interface FilmService {
 
     void removeFilmById(Long filmId);
 
-    List<Film> getPopular(Long count, int genreId, Integer year);
+    List<Film> getPopular(Long count, Long genreId, Integer year);
 
     List<Film> getFilmsSorted(int directorId, String sortBy);
